+++ .gitignore
@@ -162,3 +162,4 @@ cython_debug/
 #.idea/
 
 *.db
+.vscode
