+++ routes/exchanges.py
@@ -1,0 +1,6 @@
+from flask import jsonify
+import ccxt
+
+
+def get_exchanges():
+    return jsonify(ccxt.exchanges)
