+++ utils/db.py
@@ -1,0 +1,7 @@
+import sqlite3
+
+
+def get_db_connection():
+    conn = sqlite3.connect("kairon.db")
+    conn.row_factory = sqlite3.Row
+    return conn
